+++ pumpkin-lib/src/basic_types/stopwatch.rs
@@ -22,6 +22,10 @@ impl Stopwatch {
         self.time_start.elapsed().as_secs()
     }
 
+    pub fn get_elapsed_time_millis(&self) -> u128 {
+        self.time_start.elapsed().as_millis()
+    }
+
     pub fn get_remaining_time_budget(&self) -> i64 {
         self.time_limit_in_seconds - self.get_elapsed_time() as i64
     }